+++ commands/chat.openbsd.go
@@ -27,7 +27,6 @@ package commands
 import (
 	"bufio"
 	"fmt"
-	"os"
 	"strconv"
 	"strings"
 	"time"
@@ -118,7 +117,7 @@ func Init_Chat_Command(parentCmd *cobra.Command, app *types.AppContext) {
 				printInitialScreen()
 			}
 
-			reader := bufio.NewReader(os.Stdin)
+			reader := bufio.NewReader(app.In)
 
 			for {
 				fmt.Printf(
@@ -241,7 +240,7 @@ func Init_Chat_Command(parentCmd *cobra.Command, app *types.AppContext) {
 				if err == nil {
 					addInputToHistory(userInput)
 
-					err := quick.Highlight(os.Stdout, answer, "markdown", consoleFormatter, consoleStyle)
+					err := quick.Highlight(app.Out, answer, "markdown", consoleFormatter, consoleStyle)
 					if err != nil {
 						fmt.Print(answer)
 					}