+++ commands/prompt.go
@@ -33,7 +33,6 @@ import (
 
 func Init_Prompt_Command(parentCmd *cobra.Command, app *types.AppContext) {
 	var assistantMessages []string
-	var customModel string
 	var customTemperature float32
 	var isChat bool
 	var userMessages []string
@@ -60,7 +59,7 @@ func Init_Prompt_Command(parentCmd *cobra.Command, app *types.AppContext) {
 				systemPrompt = app.GetSystemAIPrompt("")
 			}
 
-			model := strings.TrimSpace(customModel)
+			model := strings.TrimSpace(app.Model)
 			if model == "" {
 				model = utils.GetDefaultAIChatModel()
 			}
@@ -129,7 +128,6 @@ func Init_Prompt_Command(parentCmd *cobra.Command, app *types.AppContext) {
 
 	promptCmd.Flags().StringArrayVarP(&assistantMessages, "assistant", "", []string{}, "assistant messages")
 	promptCmd.Flags().BoolVarP(&isChat, "chat", "", false, "is chat conversation and no completion operation")
-	promptCmd.Flags().StringVarP(&customModel, "model", "", "", "custom model")
 	promptCmd.Flags().Float32VarP(&customTemperature, "temperature", "", -1, "custom temperature value")
 	promptCmd.Flags().StringArrayVarP(&userMessages, "user", "", []string{}, "user messages")