+++ types/open_ai_chat.go
@@ -54,6 +54,133 @@ func (c *OpenAIChat) ClearHistory() {
 	c.Conversation = []OpenAIChatMessage{}
 }
 
+func (c *OpenAIChat) DescribeImage(message string, dataURI string) (DescribeImageResponse, error) {
+	var imageDescription DescribeImageResponse
+
+	apiKey := strings.TrimSpace(c.ApiKey)
+	if apiKey == "" {
+		return imageDescription, fmt.Errorf("no OpenAI api key defined")
+	}
+
+	model := strings.TrimSpace(strings.ToLower(c.Model))
+	if model == "" {
+		return imageDescription, fmt.Errorf("no chat ai model defined")
+	}
+
+	url := "https://api.openai.com/v1/chat/completions"
+
+	messages := []map[string]interface{}{}
+
+	// system prompt
+	if c.SystemPrompt != "" {
+		messages = append(messages, map[string]interface{}{
+			"role":    "system",
+			"content": c.SystemPrompt,
+		})
+	}
+
+	userContents := make([]map[string]interface{}, 0)
+	userContents = append(userContents, map[string]interface{}{
+		"type": "text",
+		"text": message,
+	})
+	userContents = append(userContents, map[string]interface{}{
+		"type": "image_url",
+		"image_url": map[string]interface{}{
+			"url": dataURI,
+		},
+	})
+
+	messages = append(messages, map[string]interface{}{
+		"role":    "user",
+		"content": userContents,
+	})
+
+	body := map[string]interface{}{
+		"model":    c.GetModel(),
+		"messages": messages,
+		"response_format": map[string]interface{}{
+			"type": "json_schema",
+			"json_schema": map[string]interface{}{
+				"name": "JSONAriaSchema",
+				"schema": map[string]interface{}{
+					"type":     "object",
+					"required": []string{"aria_attributes"},
+					"properties": map[string]interface{}{
+						"aria_attributes": map[string]interface{}{
+							"description": "HTML accessibility attributes which describe the image.",
+							"type":        "object",
+							"required":    []string{"aria_description", "aria_label"},
+							"properties": map[string]interface{}{
+								"aria_description": map[string]interface{}{
+									"description": "Defines a string value that describes or annotates the image in detail.",
+									"type":        "string",
+								},
+								"aria_label": map[string]interface{}{
+									"description": "Defines a string value that can be used to name the image.",
+									"type":        "string",
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+		"stream":      false,
+		"temperature": c.Temperature,
+	}
+
+	jsonData, err := json.Marshal(&body)
+	if err != nil {
+		return imageDescription, err
+	}
+
+	req, err := http.NewRequest("POST", url, bytes.NewBuffer([]byte(jsonData)))
+	if err != nil {
+		return imageDescription, err
+	}
+
+	// setup ...
+	req.Header.Set("Content-Type", "application/json")
+	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", apiKey))
+	// ... and finally send the JSON data
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		return imageDescription, err
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode != 200 {
+		return imageDescription, fmt.Errorf("unexpected response %v", resp.StatusCode)
+	}
+
+	// load the response
+	responseData, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return imageDescription, err
+	}
+
+	var chatResponse OpenAIChatCompletionResponseV1
+	err = json.Unmarshal(responseData, &chatResponse)
+	if err != nil {
+		return imageDescription, err
+	}
+
+	assistantMessage := OpenAIChatMessage{
+		Content: "",
+		Role:    "assistant",
+	}
+	if len(chatResponse.Choices) > 0 {
+		assistantMessage.Content = chatResponse.Choices[0].Message.Content
+		assistantMessage.Role = chatResponse.Choices[0].Message.Role
+	}
+
+	c.TotalTokens += chatResponse.Usage.TotalTokens
+
+	return get_ai_image_description_from_json(assistantMessage.Content)
+}
+
 func (c *OpenAIChat) GetModel() string {
 	return c.Model
 }